Move everything to the same file
- No need to separate things
